Update to the latest RandomX code

+++ src/crypto/randomx/bytecode_machine.hpp
@@ -28,10 +28,10 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #pragma once
 
-#include "common.hpp"
-#include "intrin_portable.h"
-#include "instruction.hpp"
-#include "program.hpp"
+#include "crypto/randomx/common.hpp"
+#include "crypto/randomx/intrin_portable.h"
+#include "crypto/randomx/instruction.hpp"
+#include "crypto/randomx/program.hpp"
 
 namespace randomx {
 
@@ -90,7 +90,7 @@ namespace randomx {
 		}
 
 		static void executeBytecode(InstructionByteCode* bytecode, uint8_t* scratchpad, ProgramConfiguration& config) {
-			for (int pc = 0; pc < RandomX_CurrentConfig.ProgramSize; ++pc) {
+			for (int pc = 0; pc < static_cast<int>(RandomX_CurrentConfig.ProgramSize); ++pc) {
 				auto& ibc = bytecode[pc];
 				executeInstruction(ibc, pc, scratchpad, config);
 			}
